refactor: retrofit 관련 파일 패키지 위치 변경 및 통폐합 (#19)
