--- data/src/main/java/com/yapp/breake/data/api/LoginApi.kt
@@ -1,14 +1,0 @@
-package com.yapp.breake.data.api
-
-import com.skydoves.sandwich.ApiResponse
-import com.yapp.breake.data.api.model.LoginRequest
-import com.yapp.breake.data.api.model.LoginResponse
-import retrofit2.http.Body
-import retrofit2.http.POST
-
-internal interface LoginApi {
-	@POST("/v1/auth/login")
-	suspend fun getTokens(
-		@Body request: LoginRequest,
-	): ApiResponse<LoginResponse>
-}
